refactor: split StarshipLocalHook into _af2/_af3 submodules
Match the compat-layer split pattern used elsewhere in the codebase and
let the AF3 hook subclass StarshipHttpHook so it inherits every get_*
method + set_dag_is_paused with no duplicated bodies.

* hooks/_local_af2.py  -- direct-DB LocalHook (BaseHook + StarshipHook)
* hooks/_local_af3.py  -- HTTP LocalHook (StarshipHttpHook subclass)
* hooks/starship.py    -- StarshipHook abstract + StarshipHttpHook +
                          tail dispatch to the right submodule

Also strip verbose comments in pyproject.toml.
