+++ pyproject.toml
@@ -15,9 +15,6 @@ readme = "README.md"
 license = { text = "PROPRIETARY" }
 requires-python = ">=3.6"
 dependencies = [
-    # Starship's operator-based migration flow talks HTTP end-to-end (both hooks
-    # subclass HttpHook on Airflow 3, and the target hook always does), so the
-    # HTTP provider is required at runtime, not optional.
     "apache-airflow-providers-http",
 ]
 classifiers = [
@@ -67,9 +64,6 @@ exclude = [
 ]
 
 [project.optional-dependencies]
-# Kept as an alias for backwards compatibility with older docs / install
-# instructions; the HTTP provider is now installed by default via `dependencies`
-# above.
 provider = [
     "apache-airflow-providers-http",
 ]
